+++ src/Phan/Language/Type/LiteralIntType.php
@@ -146,6 +146,17 @@ final class LiteralIntType extends IntType implements LiteralTypeInterface
             $is_nullable
         );
     }
+
+    /**
+     * Check if this type can satisfy a comparison (<, <=, >, >=)
+     * @param int|string|float|bool|null $scalar
+     * @param int $flags (e.g. \ast\flags\BINARY_IS_SMALLER)
+     * @internal
+     */
+    public function canSatisfyComparison($scalar, int $flags) : bool
+    {
+        return self::performComparison($this->value, $scalar, $flags);
+    }
 }
 
 LiteralIntType::init();