+++ tests/Phan/PHP73Test.php
@@ -37,6 +37,7 @@ final class PHP73Test extends AbstractPhanFileTest
     public function testFiles($test_file_list, $expected_file_path, $config_file_path = null)
     {
         $skip_reason = null;
+        // @phan-suppress-next-line PhanPossiblyFalseTypeArgumentInternal
         $main_path = basename(reset($test_file_list));
         if (PHP_VERSION_ID < 70300) {
             $skip_reason = 'Skip PHP 7.3 is required';